refactor(#202): cleanup UI post-#199 — split file monolitici, token CSS, err typing
- PipelineConfig.tsx: 1523 → 578 righe; estratti PersonaEditor,
  AuditPromptEditor, SettingsTabPanel, TranslationTabPanel, AuditTabPanel
  come file separati; SectionLabel sostituisce i pattern manuali;
  useEffect with eslint-disable comment per dipendenza intenzionale
- ImportPreviewDialog.tsx: 1062 → 714 righe; estratti ChunkCard,
  BoundaryDivider, SegmentEditor in ChunkEditor.tsx; utility pure
  (toParagraphChunks, countWords, toFlatModel, fromFlatModel) in
  src/utils/paragraphChunks.ts
- index.css: aggiunti token --color-editorial-paper, -page, -divider
- DocumentView.tsx + PipelineSidebar.tsx: sostituiti 7 colori hex
  hardcoded (#f4efe5, #fffdf9, #d8cfbf, #ede4d6) con token CSS;
  prerequisito per dark mode
- Header.tsx: 3x err:any → err:unknown con narrowing esplicito

Closes #202
