+++ src/components/pipeline/SettingsTabPanel.tsx
@@ -1,0 +1,191 @@
+import { ArrowRightLeft, FileText, Globe, KeyRound, Languages, Layers, ShieldCheck, Wand2 } from 'lucide-react';
+import type { Dispatch, SetStateAction } from 'react';
+import { useTranslation } from 'react-i18next';
+import type { PipelineConfig, PipelineMode, PromptTemplate } from '../../types';
+import { LANGUAGES } from '../../constants';
+import { IconButton, SectionLabel } from '../ui';
+import { PersonaEditor } from './PersonaEditor';
+
+interface SettingsTabPanelProps {
+  config: PipelineConfig;
+  setConfig: Dispatch<SetStateAction<PipelineConfig>>;
+  setMode: (mode: PipelineMode) => void;
+  translationsExist: boolean;
+  isProcessing: boolean;
+  personaTemplates: PromptTemplate[];
+  isRefiningPersona: boolean;
+  canRefinePersona: boolean;
+  personaRefineLabel: string;
+  handleRefinePersona: () => void;
+  saveTemplate: (name: string, prompt: string, context: 'stage' | 'audit' | 'persona') => Promise<void>;
+  deleteTemplate: (id: string) => Promise<void>;
+  keyStatusLoading: boolean;
+  missingRefineProviders: string[];
+}
+
+export function SettingsTabPanel({
+  config,
+  setConfig,
+  setMode,
+  translationsExist,
+  isProcessing,
+  personaTemplates,
+  isRefiningPersona,
+  canRefinePersona,
+  personaRefineLabel,
+  handleRefinePersona,
+  saveTemplate,
+  deleteTemplate,
+  keyStatusLoading,
+  missingRefineProviders,
+}: SettingsTabPanelProps) {
+  const { t } = useTranslation();
+
+  return (
+    <div id="pconfig-panel-settings" role="tabpanel" aria-labelledby="pconfig-tab-settings" className="space-y-6">
+      {/* Mode selector */}
+      <div className="space-y-2">
+        <SectionLabel icon={Layers} label={t('pipeline.modeLabel')} />
+        <div className="flex gap-2">
+          {([
+            { mode: 'standard' as PipelineMode, Icon: Languages },
+            { mode: 'editorial' as PipelineMode, Icon: Layers },
+          ]).map(({ mode: m, Icon }) => {
+            const isActive = (config.mode ?? 'standard') === m;
+            return (
+              <IconButton
+                key={m}
+                size="lg"
+                tone={isActive ? 'accent' : 'default'}
+                onClick={() => setMode(m)}
+                disabled={translationsExist || isProcessing}
+                title={t(`pipeline.mode.${m}`)}
+                ariaPressed={isActive}
+              >
+                <Icon size={16} />
+              </IconButton>
+            );
+          })}
+        </div>
+        <div className="rounded-[14px] border border-editorial-border/40 bg-editorial-textbox/20 px-3 py-3 space-y-2.5">
+          {([
+            {
+              mode: 'standard' as PipelineMode,
+              stages: [
+                { role: 'translation', Icon: Languages, labelKey: 'pipeline.stageRole.translation' },
+                { role: 'audit', Icon: ShieldCheck, labelKey: 'pipeline.tabAudit' },
+              ],
+            },
+            {
+              mode: 'editorial' as PipelineMode,
+              stages: [
+                { role: 'translation', Icon: Languages, labelKey: 'pipeline.stageRole.translation' },
+                { role: 'refine', Icon: Wand2, labelKey: 'pipeline.stageRole.refine' },
+                { role: 'format', Icon: FileText, labelKey: 'pipeline.stageRole.format' },
+                { role: 'audit', Icon: ShieldCheck, labelKey: 'pipeline.tabAudit' },
+              ],
+            },
+          ]).map(({ mode: m, stages }) => {
+            const isActive = (config.mode ?? 'standard') === m;
+            return (
+              <div key={m} className={`flex items-center gap-2.5 transition-opacity ${isActive ? '' : 'opacity-25'}`}>
+                <span className={`shrink-0 w-[68px] text-[10px] font-bold uppercase tracking-widest ${isActive ? 'text-editorial-accent' : 'text-editorial-muted'}`}>
+                  {t(`pipeline.mode.${m}`)}
+                </span>
+                <div className="flex items-center gap-1.5">
+                  {stages.map(({ role, Icon, labelKey }, i) => (
+                    <span key={role} className="flex items-center gap-1.5">
+                      {i > 0 && <span className="text-editorial-muted/40 text-xs">›</span>}
+                      <span
+                        title={t(labelKey)}
+                        aria-label={t(labelKey)}
+                        className={`inline-flex h-8 w-8 items-center justify-center rounded-full border transition-colors ${isActive ? 'border-editorial-success/40 bg-editorial-success/12 text-editorial-success' : 'border-editorial-border bg-editorial-bg text-editorial-muted'}`}
+                      >
+                        <Icon size={14} strokeWidth={1.9} />
+                      </span>
+                    </span>
+                  ))}
+                </div>
+              </div>
+            );
+          })}
+        </div>
+      </div>
+
+      {/* Language pair */}
+      <div className="space-y-2">
+        <SectionLabel icon={Globe} label={t('pipeline.languagePair')} />
+        <div className={`flex items-center gap-3 transition-opacity ${!!config.persona ? 'opacity-40 pointer-events-none' : ''}`}>
+          <select
+            value={config.sourceLanguage}
+            onChange={(e) => setConfig((prev) => ({ ...prev, sourceLanguage: e.target.value }))}
+            className="w-full rounded-[14px] border border-editorial-border bg-editorial-bg/80 px-3 py-2 text-xs font-mono outline-none focus-visible:ring-2 focus-visible:ring-editorial-accent appearance-none"
+            aria-label={t('pipeline.sourceLanguage')}
+            disabled={!!config.persona}
+          >
+            {LANGUAGES.map((lang) => (
+              <option key={lang} value={lang}>{t(`languages.${lang}`)}</option>
+            ))}
+          </select>
+          <IconButton
+            size="md"
+            className="shrink-0"
+            onClick={() =>
+              setConfig((prev) => ({
+                ...prev,
+                sourceLanguage: prev.targetLanguage,
+                targetLanguage: prev.sourceLanguage,
+              }))
+            }
+            disabled={!!config.persona}
+            title={t('pipeline.swapLanguages')}
+          >
+            <ArrowRightLeft size={13} />
+          </IconButton>
+          <select
+            value={config.targetLanguage}
+            onChange={(e) => setConfig((prev) => ({ ...prev, targetLanguage: e.target.value }))}
+            className="w-full rounded-[14px] border border-editorial-border bg-editorial-bg/80 px-3 py-2 text-xs font-mono outline-none focus-visible:ring-2 focus-visible:ring-editorial-accent appearance-none"
+            aria-label={t('pipeline.targetLanguage')}
+            disabled={!!config.persona}
+          >
+            {LANGUAGES.map((lang) => (
+              <option key={lang} value={lang}>{t(`languages.${lang}`)}</option>
+            ))}
+          </select>
+        </div>
+        {!!config.persona && (
+          <p className="text-xs leading-relaxed text-editorial-muted/60">
+            {t('pipeline.languagePairLockedByPersona')}
+          </p>
+        )}
+      </div>
+
+      <PersonaEditor
+        persona={config.persona}
+        sourceLanguage={config.sourceLanguage}
+        targetLanguage={config.targetLanguage}
+        templates={personaTemplates}
+        isRefining={isRefiningPersona}
+        canRefine={canRefinePersona}
+        refineLabel={personaRefineLabel}
+        onChange={(value) => setConfig((prev) => ({ ...prev, persona: value }))}
+        onRefine={handleRefinePersona}
+        onSaveTemplate={(name, prompt) => saveTemplate(name, prompt, 'persona')}
+        deleteTemplate={deleteTemplate}
+      />
+
+      {/* Refine keys status */}
+      <div className="space-y-2">
+        <SectionLabel icon={KeyRound} label={t('pipeline.refineKeyLabel')} />
+        {!keyStatusLoading && (
+          <p className="text-[11px] leading-relaxed text-editorial-muted">
+            {missingRefineProviders.length > 0
+              ? t('pipeline.refineKeyMissingHint', { providers: missingRefineProviders.join(', ') })
+              : t('pipeline.refineKeyReadyHint')}
+          </p>
+        )}
+      </div>
+    </div>
+  );
+}
